requirements.txt
discord.py
pyowm
requests